componentize session modals
